# docs/swagger.yaml
basePath: /fab
definitions:
...
  channel.createChannelRequest:
    properties:
      channelid:
        example: mychannel
        type: string
      orderernames:
        example:
        - orderer0
        - orderer1
        - orderer2
        items:
          type: string
        type: array
      ordererorgname:
        example: orderer
        type: string
      ordererusername:
        example: Admin
        type: string
      orgmsps:
        example:
        - Org1MSP
        - Org2MSP
        - Org3MSP
        items:
          type: string
        type: array
      orgnames:
        example:
        - org1
        - org2
        - org3
        items:
          type: string
        type: array
    required:
    - orderernames
    - orgmsps
    - orgnames
    type: object
  channel.joinChannelRequest:
    properties:
      channelid:
        example: mychannel
        type: string
      orderernames:
        example:
        - orderer0
        - orderer1
        - orderer2
        items:
          type: string
        type: array
      orgmsps:
        example:
        - Org1MSP
        - Org2MSP
        - Org3MSP
        items:
          type: string
        type: array
      orgnames:
        example:
        - org1
        - org2
        - org3
        items:
          type: string
        type: array
    required:
    - orderernames
    - orgmsps
    - orgnames
    type: object
  gintool.ApiResponse:
    properties:
      code:
        description: 状态码
        type: integer
      err_msg:
        description: 内部错误详情
        type: string
      message:
        description: 状态短语
        type: string
      result:
        description: 数据结果集
        type: object
    type: object
...
host: localhost:6922
info:
  contact: {}
  description: 描述:fabapi 接口文档.
  license: {}
  title: fabapi API文档
  version: 1.1.0
paths:
...
  /channel/createchannel:
    post:
      consumes:
      - application/json
      description: 功能：创建通道管理
      parameters:
      - description: 创建通道管理
        in: body
        name: resource
        required: true
        schema:
          $ref: '#/definitions/channel.createChannelRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gintool.ApiResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gintool.ApiResponse'
            type: object
      summary: 创建通道管理
      tags:
      - 创建通道管理
  /channel/joinchannel:
    post:
      consumes:
      - application/json
      description: 功能：加入通道管理
      parameters:
      - description: 加入通道管理
        in: body
        name: resource
        required: true
        schema:
          $ref: '#/definitions/channel.joinChannelRequest'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gintool.ApiResponse'
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/gintool.ApiResponse'
            type: object
      summary: 加入通道管理
      tags:
      - 加入通道管理
...
swagger: "2.0"